[dataset] remove network dataset buffers
This commit removes both the Active and Pending Operational Dataset buffers.
