+++ src/core/meshcop/dataset.cpp
@@ -32,12 +32,15 @@
  *
  */
 
+#define WPP_NAME "dataset.tmh"
+
 #include "dataset.hpp"
 
 #include <stdio.h>
 
 #include "common/code_utils.hpp"
 #include "common/instance.hpp"
+#include "common/logging.hpp"
 #include "common/settings.hpp"
 #include "meshcop/meshcop_tlvs.hpp"
 #include "thread/mle_tlvs.hpp"
@@ -528,6 +531,137 @@ void Dataset::Remove(uint8_t *aStart, uint8_t aLength)
 {
     memmove(aStart, aStart + aLength, mLength - (static_cast<uint8_t>(aStart - mTlvs) + aLength));
     mLength -= aLength;
+}
+
+otError Dataset::ApplyConfiguration(Instance &aInstance) const
+{
+    ThreadNetif &netif = aInstance.GetThreadNetif();
+    Notifier &notifier = aInstance.GetNotifier();
+    Mac::Mac &mac = netif.GetMac();
+    otError error = OT_ERROR_NONE;
+    const Tlv *cur = reinterpret_cast<const Tlv *>(mTlvs);
+    const Tlv *end = reinterpret_cast<const Tlv *>(mTlvs + mLength);
+
+    while (cur < end)
+    {
+        switch (cur->GetType())
+        {
+        case Tlv::kChannel:
+        {
+            uint8_t channel = static_cast<uint8_t>(static_cast<const ChannelTlv *>(cur)->GetChannel());
+
+            if (mac.GetChannel() != channel)
+            {
+                error = mac.SetChannel(channel);
+
+                if (error != OT_ERROR_NONE)
+                {
+                    otLogWarnMeshCoP(aInstance,
+                                     "DatasetManager::ApplyConfiguration() Failed to set channel to %d (%s)",
+                                     channel, otThreadErrorToString(error));
+                    ExitNow();
+                }
+
+                notifier.SetFlags(OT_CHANGED_THREAD_CHANNEL);
+            }
+
+            break;
+        }
+
+        case Tlv::kPanId:
+        {
+            uint16_t panid = static_cast<const PanIdTlv *>(cur)->GetPanId();
+
+            if (mac.GetPanId() != panid)
+            {
+                mac.SetPanId(panid);
+                notifier.SetFlags(OT_CHANGED_THREAD_PANID);
+            }
+
+            break;
+        }
+
+        case Tlv::kExtendedPanId:
+        {
+            const ExtendedPanIdTlv *extpanid = static_cast<const ExtendedPanIdTlv *>(cur);
+
+            if (memcmp(mac.GetExtendedPanId(), extpanid->GetExtendedPanId(), OT_EXT_PAN_ID_SIZE) != 0)
+            {
+                mac.SetExtendedPanId(extpanid->GetExtendedPanId());
+                notifier.SetFlags(OT_CHANGED_THREAD_EXT_PANID);
+            }
+
+            break;
+        }
+
+        case Tlv::kNetworkName:
+        {
+            const NetworkNameTlv *name = static_cast<const NetworkNameTlv *>(cur);
+            otNetworkName networkName;
+            memcpy(networkName.m8, name->GetNetworkName(), name->GetLength());
+            networkName.m8[name->GetLength()] = '\0';
+
+            if (strcmp(networkName.m8, mac.GetNetworkName()) != 0)
+            {
+                mac.SetNetworkName(networkName.m8);
+                notifier.SetFlags(OT_CHANGED_THREAD_NETWORK_NAME);
+            }
+
+            break;
+        }
+
+        case Tlv::kNetworkMasterKey:
+        {
+            const NetworkMasterKeyTlv *key = static_cast<const NetworkMasterKeyTlv *>(cur);
+            netif.GetKeyManager().SetMasterKey(key->GetNetworkMasterKey());
+            break;
+        }
+
+#if OPENTHREAD_FTD
+
+        case Tlv::kPSKc:
+        {
+            const PSKcTlv *pskc = static_cast<const PSKcTlv *>(cur);
+            netif.GetKeyManager().SetPSKc(pskc->GetPSKc());
+            break;
+        }
+
+#endif
+
+        case Tlv::kMeshLocalPrefix:
+        {
+            const MeshLocalPrefixTlv *prefix = static_cast<const MeshLocalPrefixTlv *>(cur);
+            netif.GetMle().SetMeshLocalPrefix(prefix->GetMeshLocalPrefix());
+            break;
+        }
+
+        case Tlv::kSecurityPolicy:
+        {
+            const SecurityPolicyTlv *securityPolicy = static_cast<const SecurityPolicyTlv *>(cur);
+            netif.GetKeyManager().SetKeyRotation(securityPolicy->GetRotationTime());
+            netif.GetKeyManager().SetSecurityPolicyFlags(securityPolicy->GetFlags());
+            break;
+        }
+
+        default:
+        {
+            break;
+        }
+        }
+
+        cur = cur->GetNext();
+    }
+
+exit:
+    return error;
+}
+
+otError Dataset::ConvertToActive(void)
+{
+    Remove(Tlv::kPendingTimestamp);
+    Remove(Tlv::kDelayTimer);
+    mType = Tlv::kActiveTimestamp;
+    return OT_ERROR_NONE;
 }
 
 }  // namespace MeshCoP
